[DOCS,REFACTORING] : 주석 추가와 패키지 구조 변경, Router가 시작점이 가지지 않고, RouteTable이 가지도록 변경
