--- backend/src/main/java/com/almostThere/domain/map/Service/router/Router.java
@@ -1,6 +1,0 @@
-package com.almostThere.domain.map.Service.router;
-
-public interface Router {
-    RouteInfo[] getShortestPath();
-    public void showPath(Long map_dest_id);
-}
