--- backend/src/main/java/com/almostThere/domain/map/Service/test/context/Context.java
@@ -1,6 +1,0 @@
-package com.almostThere.domain.map.Service.test.context;
-
-public interface Context {
-    void action();
-    void setState(String state);
-}
