--- backend/src/main/java/com/almostThere/domain/map/Service/test/state/TestState.java
@@ -1,5 +1,0 @@
-package com.almostThere.domain.map.Service.test.state;
-
-public interface TestState {
-    void action();
-}
