--- backend/src/main/java/com/almostThere/domain/map/entity/node/MapNode.java
@@ -1,22 +1,0 @@
-package com.almostThere.domain.map.entity.node;
-
-import lombok.Getter;
-
-@Getter
-public class MapNode {
-    long map_id;
-    double longitude;
-    double latitude;
-
-    String name;
-
-    public MapNode(long map_id, double longitude, double latitude, String name) {
-        this.map_id = map_id;
-        this.longitude = longitude;
-        this.latitude = latitude;
-        this.name = name;
-        if (name == null) {
-            this.name = "walk";
-        }
-    }
-}
