--- backend/src/main/java/com/almostThere/domain/map/repository/graphBuilder/MapGraphBuilder.java
@@ -1,10 +1,0 @@
-package com.almostThere.domain.map.repository.graphBuilder;
-
-import com.almostThere.domain.map.entity.link.MapLink;
-import com.almostThere.domain.map.entity.node.MapNode;
-import com.almostThere.domain.map.repository.mapGraph.MapGraph;
-import java.util.List;
-
-public interface MapGraphBuilder {
-    MapGraph build(List<MapNode> nodes, List<MapLink> links);
-}
