+++ backend/src/main/java/com/almostThere/middleSpace/domain/gis/Position.java
@@ -1,0 +1,11 @@
+package com.almostThere.middleSpace.domain.gis;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class Position {
+    private Double latitude;
+    private Double longitude;
+}
