+++ backend/src/main/java/com/almostThere/middleSpace/service/routing/Router.java
@@ -1,0 +1,7 @@
+package com.almostThere.middleSpace.service.routing;
+
+import com.almostThere.middleSpace.domain.routetable.RouteTable;
+
+public interface Router {
+    RouteTable getShortestPath(int startNodeIndex);
+}
