+++ backend/src/main/java/com/almostThere/middleSpace/test/context/Context.java
@@ -1,0 +1,6 @@
+package com.almostThere.middleSpace.test.context;
+
+public interface Context {
+    void action();
+    void setState(String state);
+}
