+++ backend/src/main/java/com/almostThere/middleSpace/test/state/TestState.java
@@ -1,0 +1,5 @@
+package com.almostThere.middleSpace.test.state;
+
+public interface TestState {
+    void action();
+}
